backend/pyproject.toml
[tool.poetry]
name = "adrenaline"
version = "0.1.0"
description = "A helpful clinical AI assistant"
authors = ["Vector Institute <ai_engineering@vectorinstitute.ai>"]
repository = "https://github.com/VectorInstitute/adrenaline"
license = "Apache-2.0"
packages = [
    { include = "api/" },
]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11, <3.12"
fastapi = "^0.111.0"
pydantic = "^1.10.11"
uvicorn = "^0.30.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "3.2.0"
passlib = "^1.7.4"
pyjwt = "^2.9.0"
sqlalchemy = { version = "^2.0.32", extras = ["asyncio"] }
aiosqlite = "^0.20.0"
motor = "^3.5.1"
medcat = "^1.13.0"
polars-lts-cpu = "^1.7.0"

[tool.poetry.group.jupyterlab]
optional = true

[tool.poetry.group.jupyterlab.dependencies]
jupyterlab = "^4.2.5"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^7.1.1"
pre-commit = "^2.17.0"
pytest-cov = "^3.0.0"
codecov = "^2.1.13"
mypy = "^1.7.0"
ruff = "^0.6.0"
pip-audit = "^2.7.1"

[tool.mypy]
plugins = ["pydantic.mypy"]
ignore_missing_imports = true
install_types = true
pretty = true
namespace_packages = true
explicit_package_bases = true
non_interactive = true
warn_unused_configs = true
allow_any_generics = false
allow_subclassing_any = false
allow_untyped_calls = false
allow_untyped_defs = false
allow_incomplete_defs = false
check_untyped_defs = true
allow_untyped_decorators = false
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
implicit_reexport = false
strict_equality = true
extra_checks = true

[tool.ruff]
include = ["*.py", "pyproject.toml", "*.ipynb"]
line-length = 88

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint]
select = [
    "A", # flake8-builtins
    "B", # flake8-bugbear
    "COM", # flake8-commas
    "C4", # flake8-comprehensions
    "RET", # flake8-return
    "SIM", # flake8-simplify
    "ICN", # flake8-import-conventions
    "Q", # flake8-quotes
    "RSE", # flake8-raise
    "D", # pydocstyle
    "E", # pycodestyle
    "F", # pyflakes
    "I", # isort
    "W", # pycodestyle
    "N", # pep8-naming
    "ERA", # eradicate
    "PL", # pylint
]
fixable = ["A", "B", "COM", "C4", "RET", "SIM", "ICN", "Q", "RSE", "D", "E", "F", "I", "W", "N", "ERA", "PL"]
ignore = [
    "B905", # `zip()` without an explicit `strict=` parameter
    "E501", # line too long
    "D203", # 1 blank line required before class docstring
    "D213", # Multi-line docstring summary should start at the second line
    "PLR2004", # Replace magic number with named constant
    "PLR0913", # Too many arguments
    "COM812", # Missing trailing comma
]

# Ignore import violations in all `__init__.py` files.
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401", "F403", "F811"]

[tool.ruff.lint.pep8-naming]
ignore-names = ["X*", "setUp"]

[tool.ruff.lint.isort]
lines-after-imports = 2

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.pycodestyle]
max-doc-length = 88

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
